Server: Fix BadMatch when resizing w/ X RandR 1.2+
Referring to the commit message for
fbc2f420cdad982140963a82f62f1756b7b23bca, I asserted that "The TigerVNC
equivalent [of the vncRRCrtcSet() function] twiddles the output connection
state and calls RRCrtcNotify(), but in my testing, I was never able to
find a situation in TurboVNC under which that was necessary."

Well, I found the situation.  :|  When xrandr is invoked with -s, it
uses the X RandR 1.0 API.  This causes ProcRRSetScreenConfig() to be
called on the server.  That function calls RRCrtcSet() to set the modes
for all of the screen's CRTCs to NULL, then it calls RRScreenSizeSet(),
then it calls RRCrtcSet() again to update the mode of the screen's
first CRTC.  For that use case, it's OK for our implementation of
RRCrtcSet() to always return TRUE, because RRCrtcNotify() is called
within the body of our implementation of RRScreenSizeSet().  However,
when xrandr is invoked with --output {output} --mode {new_mode}, it
uses the X RandR 1.2 API.  This can cause ProcRRSetCrtcConfig() to be
called with a NULL CRTC (which subsequently causes RRCrtcSet() to be
called with a NULL CRTC.) Unless RRCrtcNotify() is called within the
body of our RRCrtcSet() implementation, then the old mode remains
attached to the CRTC, and the check at randr/randr.c:277 fails if the
new mode has smaller dimensions than the old mode.