Un*x/Java viewer: Multi-screen spanning in FS mode
It's still not possible to implement this on Un*x in windowed mode,
because creating a multi-screen non-full-screen X window requires
forcing the minimum width/height using XSetWMNormalHints(), and this
results in a window that can't be resized.

This commit also introduces fallback code on Un*x that will use
GraphicsDevice.setFullScreenWindow() if the TurboVNC Helper isn't
available.  That will at least allow full-screen mode to work with
the primary screen without requiring JNI.