Win: Fix multi-screen span bug w/ swapped screens
Due to some vestigial code left over from
5dd4d22aa2b4467dd16f26fd90724bbf0dcf8e5f, the non-full-screen work area
was never being expanded onto screens that were to the left of or above
the primary.

This also fixes a subtle error in the computation of the multi-screen
areas.  Although unlikely, it is possible that two vertically offset
screens could both overlap vertically with the primary but not with each
other.  Thus, we should compare the vertical overlap of each screen
against the span area computed so far.