Server: Clamp desktop dimensions to 32767x32767
The underlying X.org screen structure uses a signed short to represent
width and height, so values > 32767 overflow the data type and can be
interpreted as negative numbers.  If a width/height larger than 32767
was passed in the server's -geometry argument or received from a
viewer in an ExtendedDesktopSize request, the server would subsequently
fail to send any framebuffer updates and would continuously log
messages of the form:

WARNING: Framebuffer update at 0,0 with dimensions 0x0 has been clipped
to the screen boundaries

The RFB server was still working properly, even though the X server
wasn't, so resetting the desktop size using xrandr or an
ExtendedDesktopSize request would fix this problem.  However, if a user
attempted to configure and enable a new X RandR mode with width or
height > 32767, the server segfaulted.

The server will now silently clamp any requested desktop dimensions to
32767x32767, regardless of the value of the max-desktop-size directive
in turbovncserver-security.conf.

This commit also closes a loophole that allowed users to exceed the
dimensions specified in the max-desktop-size security configuration file
directive by using X RandR functions or the xrandr utility.