Viewer: "auto" resize+span= "Primary" win,"All" FS
The Automatic spanning mode wasn't really useful when automatic desktop
resizing was enabled, because it did nothing differently than Primary
spanning.  This commit changes the behavior of the auto resize+span mode
such that it behaves similarly to the TigerVNC Viewer, using
multi-screen spanning in full-screen mode and primary spanning in
windowed mode.

This commit also clarifies and documents the spanning logic,
particularly in the Java viewer.