Win/Java: Fix failure to launch w/ 32-bit JRE
Referring to:
http://www.oracle.com/technetwork/java/hotspotfaq-138619.html#compiler_types
32-bit JREs on Windows only provide the client JVM, and 64-bit JREs only
provide the server JVM, so specifying -server in vncviewer-java[w].bat
was causing a JVM error ("Error: missing 'server' JVM".)  With Java Web
Start, the error was silent, so the TurboVNC Viewer simply did not
launch.

NOTE: -server is never relevant with 64-bit JREs, but OS X and Linux
32-bit JREs will only enable the server JVM by default if they detect a
"server-class" machine, so we continue to specify -server for those
platforms.  However, the documentation section describing -server
has been removed, since there is very little reason why a user would
ever need to specify that anymore (that section was written before the
Java TurboVNC Viewer was packaged as a standalone application.)