Java: Support a few more .vnc file directives
The Windows viewer recognizes the following directives in connection
info files:

"user" (which can be used to enable Unix Login or Plain authentication),

"noremotecursor" (which can be used to disable both cursor shape updates
and server-side cursor rendering), and

"restricted" (which disables Ctrl-Alt-Del and Ctrl-Esc.)

Since the Java viewer now supports those options as well, this commit
maps the directives to viewer parameters accordingly.

Note that "restricted" is completely undocumented, which is why the
addition of that directive is not mentioned in the change log.  That
directive is solely for the use of SysAdmins and portal designers.