Win: Ensure complex RFB msgs are sent atomically
It is possible for a pointer or keyboard event to occur between calls
to WriteExact(), and because our Windows viewer sends EDS, GII, and
ClientCutText messages using multiple WriteExact() calls, we need to
lock the write mutex before the first call in the sequence to ensure
that the entire message is sent as a unit.  Otherwise, stray input
events could corrupt the message.

It is unknown whether this fixes a specific user-visible bug, but there
were vague reports about occasional failures to send large amounts of
clipboard data with the Windows viewer, and it's possible that this was
the cause.