Java: Check clipboard when mouse clicked in window
If the viewer window doesn't have input focus, then clicking a mouse
button in the window will bring it into focus on Windows and Un*x
platforms, but the focus listener for the child window (DesktopWindow)
will be invoked after the mouse listener.  Thus, if text is selected in
the TurboVNC Server session, the client's new clipboard contents won't
replace the server's unless we explicitly check the clipboard in the
mouse listener as well as the focus listener.

NOTES:
- Java unfortunately doesn't have a "clipboard changed" event, hence
  the need to check the clipboard when other types of events are
  triggered.
- On Mac platforms, clicking the middle mouse button in the window
  doesn't bring it into focus, and only the mouse press event is
  communicated to the viewer.  The clipboard is still transferred
  properly, because of this commit, but the middle mouse button is left
  in a pressed state on the server.  This seems to be an OS X bug of
  some sort, because the TigerVNC (native) Viewer suffers from it as
  well.

Fixes #105