Server: Don't build GLX extension on Mac
I was never able to make it work with the swrast_dri.so provided with
XQuartz, and it screws up the i386 build because of lack of TLS support.
This commit also removes the unnecessary TLS macro definition.  The
newer version of Mesa we're now using only supports __thread or nothing.