Server: Hopefully fix rare desktop resize segfault
I was able to reproduce the segfault reliably for about 30 minutes on a
very specific client/server configuration, but then it magically went
away.  Unfortunately the fix cannot be described in the change log
until/unless the exact cause can be nailed down, but what seemed to be
happening was that either:

(1) A framebuffer update was somehow triggered within the body of
xf86SetRootClip() without going through the normal deferred
update mechanism, so the update was not blocked.

... OR ...

(2) When rfbScreen was assigned a new value in randr.c:364, this had
the effect of resetting rfbScreen.blockUpdates prematurely, so a new
update was triggered in the subsequent call to xf86SetRootClip().

I wasn't able to reproduce the issue for long enough to figure out
exactly why the segfault occurred, but this commit attempts to shore
up the desktop resize logic in order to proactively prevent the issue
from occurring in the future.