Server: Fix VeNCrypt incompat. w/ LibVNCClient
... and possibly other VeNCrypt client implementations.  The TurboVNC
Server was incorrectly sending the VeNCrypt version as minor.Major
instead of Major.minor.  Because of a buglet in the Java viewer
(inherited from TigerVNC) whereby that viewer incorrectly receives the
version in minor.Major order, I erroneously assumed that the server
should send the version in that same order.  Both TigerVNC viewers and
our Java viewer will accept any VeNCrypt version >= 0.2, which is why
this bug went unnoticed until now, but LibVNCClient requires that the
version be exactly 0.2.

This commit also fixes the aforementioned Java viewer buglet.