Java: Fix multi-connection KB grabbing issues
A description of a Gordian knot:

This started with an observation that, when opening a new viewer
connection, the old connection window would detect that it lost focus,
and it would ungrab the keyboard.  Because ungrabbing the keyboard
applies to the whole process, this led to a race condition whereby
the keyboard was often ungrabbed in the new connection window, despite
the viewer thinking that it was grabbed.

Step 1: When a connection window loses focus, only ungrab the keyboard
if focus was lost to a non-TurboVNC window.

Easy enough, but this led to another problem: if multiple connection
windows were active and focus switched between them, the keyboard
grabbing state would remain unchanged, even if the grab mode differed
between the connections.  On Un*x/X11 systems, keyboard grabbing applies
to the whole process, so

Step 2: When a connection window regains focus, grab or ungrab the
keyboard to match the connection's chosen grab mode (if the mode is
Always), full-screen mode (if the mode is FS), or grab state (if the
mode is Manual.)

Easy enough, but this led to another problem: if Manual grab mode was
selected, there was no way to determine whether grabbing was enabled
before the connection window lost focus.

Step 3: Decouple the actual grab state from the selected grab state.

The actual grab state is now stored more robustly, in a synchronized
static variable that allows all viewer connection threads to see whether
grabbing is active on any other viewer connection.  The selected grab
state is simply read from the F8 menu object.

All of this could have been much more easily solved by forking a new
JVM for the second and subsequent connections, but that involves some
rather hackish things that are likely to interfere with Java Web Start.