Win: Disable Windows Input Method Editor system
Most Asian languages require IME in order to input language-specific
characters, but in order for this to work properly with the TurboVNC
Viewer, we would need to handle the VK_PROCESSKEY virtual key code and
use Win32 Input Method Manager functions to translate the
language-specific characters into X11 keysyms.  Barring that, the IME
does nothing except interfere with the normal operation of the viewer.

Closes #124