Fix build errors when C flags include -Werror
Instructing the compiler to treat warnings as errors caused some of the
compiler tests to fail, because the test code was not 100% clean.

Note that we now use check_symbol_exists() instead of
check_function_exists() or check_library_exists() (which is basically
just a wrapper for check_function_exists()), since the test code for
check_function_exists() produces a compiler warning due to not including
the appropriate headers.