Mac: Work around JDK-8208712
... which caused the Profiling dialog, when invoked using the Command-P
hotkey, to pop up and then immediately disappear.  The proximate cause
is that the ActionListener in MacMenuBar is invoked twice due to the
following JDK bug:

https://bugs.openjdk.java.net/browse/JDK-8208712

which affects JCheckBoxMenuItems in JMenuBars with
apple.laf.useScreenMenuBar enabled under Java 9+.  The workaround is
applied only under Java 9 through 11, since the JDK bug report states
that the issue will be fixed in Java 12.  Will re-evaluate once Java 12
is released.