Win: Don't allow 32-bit installer on Windows/x64
In addition to reduced performance, there is now a more serious drawback
(lack of a custom JRE) to installing the Windows/x86 TurboVNC Viewer
package on an x64 system.  Our download numbers suggest that a lot of
64-bit Windows users are downloading the 32-bit TurboVNC Viewer for
Windows.  This commit and 4a98f4fff27fa3b6c76d2cb97f48ea77600305f9 are
meant to discourage that behavior and to wean users off of the 32-bit
Windows packages.