Win: Get rid of our custom version of PuTTY
A custom version of PuTTY was originally included because it was
significantly faster than the stock version of PuTTY 0.60, but these
days, the stock version of PuTTY (0.70) is significantly faster than our
custom version, and numerous other Windows SSH solutions exist (Bitvise,
Win32-OpenSSH, etc.) that work properly and optimally with the TurboVNC
Viewer.  There are even two solutions (MSYS2 and Cygwin) that work
properly with the -via and -tunnel options in the native Windows
TurboVNC Viewer.  To reflect that fact, the default values of
VNC_VIA_CMD and VNC_TUNNEL_CMD in the Windows TurboVNC Viewer now refer
to ssh.exe instead of c:\cygwin\bin\ssh.exe, so the Windows viewer will
work properly with whichever of Cygwin OpenSSH or MSYS2 OpenSSH is in
the PATH.