vncserver: Don't require X11 TCP when creating OTP
Passing '-display $host:$displayNumber' to vncpasswd caused it to
(attempt to) use a TCP connection when connecting to the X11 display, if
$host didn't resolve to the loopback device.  This isn't a problem with
most systems, on which /etc/hosts is configured such that the hostname
is equivalent to localhost, but some machines map the hostname to the
external IP address in /etc/hosts.  On such machines, it is necessary to
pass '-display :$displayNumber' to vncpasswd in order to make it use a
Unix domain socket.