Server: noVNC integration
Referring to #147, this seems like the path of least resistance for the
moment.  It allows users to continue using JWS if they want, but it
provides a way forward for a Java-less zero-install VNC viewer for
TurboVNC sessions, until/unless our project has the cycles to develop
such a viewer ourselves.

(This commit also moves the old zero-install JWS documentation from the
User's Guide to the Java viewer's README file.)