Fix sess. mgr. when host doesn't know its hostname
When SessMgrAuto=true, the viewer should behave as if -tunnel was
specified, thus causing Tunnel.createTunnel() to set the TurboVNC host
to "localhost", relative to the gateway.  Failing to do this causes the
remote end of the SSH tunnel to connect to {host}:{displayNum} instead
of localhost:{displayNum}, which will fail if {host} is not in the
host's /etc/hosts file.