Server: Use system-wide xstartup.turbovnc script
... if the per-user xstartup.turbovnc script (usually under ~/.vnc)
doesn't exist.  Since xstartup.turbovnc often contains bug fixes, this
commit allows that script to be more easily updated on a system-wide
basis.  The old behavior (creating the per-user xstartup.turbovnc script
with default contents, if it didn't exist) was a relic of the days in
which xstartup launched TWM by default, so it was almost always
desirable to modify it.  These days, xstartup.turbovnc is designed to
launch numerous different window managers across a variety of platforms.
Whereas we provide several environment variables that can be used to
customize the behavior of the script, the script itself isn't really
customizable.

Note that we install xstartup.turbovnc in the same directory as
vncserver, so vncserver can find the script regardless of where TurboVNC
is installed.