Viewer: Move TurboJPEG interface into the Helper
The ability to use TurboJPEG through JNI was introduced with the
overhaul of the Java TurboVNC Viewer in TurboVNC 1.2, but the TurboVNC
Helper wasn't introduced until TurboVNC 2.0 and wasn't available on all
supported platforms until TurboVNC 2.2.  The TurboVNC build system also
had the ability to sign the TurboJPEG JNI JARs using the same
certificate that was used to sign the TurboVNC Viewer JAR, thus allowing
those JARs to be packaged with the TurboVNC Server and deployed
automatically to Linux, Mac, and Windows clients via Java Web Start
(whereas doing likewise with the TurboVNC Helper would have created a
chicken-and-egg dilemma.)

Thus, it made sense historically to use a separate JNI library for
TurboJPEG.  However, since Java Web Start is no longer a thing and the
TurboVNC Helper is built by default on all supported platforms, there is
no longer any need to use a separate JNI library for TurboJPEG.  To that
end, this commit creates a dedicated JNI wrapper for the TurboJPEG C API
and includes it in the TurboVNC Helper library.

The main purpose of this commit at the moment is to streamline the build
and allow distribution-specific builds of the TurboVNC Viewer on Linux
distributions that provide the TurboJPEG C API but not the TurboJPEG
Java API (including RHEL/CentOS 7 via EPEL, RHEL/CentOS 8 via
CodeReady/PowerTools, recent Ubuntu releases, and Fedora.)  However, it
also lays the groundwork for moving the entire Tight decoder into the
Helper, which is a long-term project goal (refer to #144.)