Build: Use install rather than cp to build Mac app
cp doesn't always honor umask, so if the TurboVNC source directory
wasn't world-readable, some of the files in the generated
TurboVNC Viewer.app bundle would not be world-readable.  The specific
problem I observed was that the app bundle icon wasn't displayed until
the app was launched.  This was due to the icons, located under
Contents/Resources in the app bundle, having permissions of 600.

This was not generally a problem when using the official build scripts,
since those scripts ensure that the TurboVNC source directory is
world-readable.