Viewer: Compare private keys more robustly
Referring to the previous commit, the name of the private key fetched
from the agent will never match the private key filename on Mac and
Windows systems, and it won't necessarily match the private key filename
on Un*x systems either (because the private key file may have been
copied.)  Instead, this commit causes JSch to match private keys using
their public key fingerprints.  Thus, JSch will now ignore a private key
specified in the OpenSSH config file or on the TurboVNC Viewer command
line unless one of the following is true:

- The corresponding public key ({private_key_filename}.pub} doesn't
  exist for the specified private key file.
- The public key fingerprint for the specified private key file doesn't
  match the public key fingerprint of any key provided by the agent.
- The SSHKeyPass parameter is specified along with SSHKeyFile.

This is essentially the same behavior as OpenSSH with
IdentitiesOnly=yes.