Viewer: Prioritize specified SSH private key ...
... over the private keys offered by the SSH agent.  Previously, the
viewer's built-in SSH client avoided adding a key that was explicitly
specified, in the OpenSSH config file or on the command line, if a
decrypted version of the same key was already being offered by the
agent.  However, this wasn't entirely correct behavior.  The OpenSSH
server is configured with MaxAuthTries=6 by default, so if the SSH agent
is offering more than 6 private keys and the correct key for a
particular server is not one of the first 6, then an error ("Too many
authentication failures") would normally occur.  The preferred way to
work around that error, when using OpenSSH, is either (a) to create an
OpenSSH config file entry for the host and specify the SSH private key
in the host entry's IdentityFile directive, or (b) to specify the SSH
private key on the command line.  However, neither method worked around
the issue when using the TurboVNC Viewer, because when the viewer
avoided adding the explicitly-specified duplicate key, the SSH agent's
key order was preserved.  The correct behavior is instead to add the
specified key, thus giving it priority, but to add the decrypted version
of it from the SSH agent rather than the encrypted version.