Server: Don't use client ptr after disconnect,Pt.3
Use the nextCl trick from TightVNC to prevent invalid reads (and
potential segfaults) if a viewer disconnects while resizing the desktop
or changing the cursor shape, and other viewers are still connected, or
if vncconnect is used to disconnect all listening viewers while a
non-listening viewer is still connected.

Also:
- Allow rfbSendDesktopSizeAll(..., rfbEDSReasonServer) to successfully
send an EDS update to viewers that remain connected, even if a viewer
disconnects while attempting to send an EDS update to it.
- Allow client messages to be successfully processed from viewers that
remain connected, even if a viewer disconnects while attempting to
process client messages from it.
- Move the code for checking the validity of a client pointer into a
new macro.
- Revert 742370cd812a7b1855b0826022c198943d6e4b0f, since it is no
longer necessary.