Java: Fix SSLHandshakeException w/ TLS* sec. types
Java 7u211, 8u201, and 11.0.2 now add the anonymous cipher suites to the
jdk.tls.disabledAlgorithms property in the java.security file, so we
need to twiddle that property programmatically in order to enable them.

(And yes, I know that anonymous cipher suites are vulnerable to
man-in-the-middle attacks, but they're also part of the VeNCrypt
specification, and they're better than no encryption.  TurboVNC supports
SSH tunneling or X.509 certificates for those who need more robust
encryption.)

Fixes #159