Server: Fix flickering w/ UltraVNC Viewer
Regression caused by b45158e4b05f2e853bb83240a4e61accf3139d27

The UltraVNC Viewer seems to not send incremental FBURs, so we need to
separate the pending resize logic for "normal" and extended desktop size
messages and trigger a pending resize for the "normal" flavor only in
response to a server-side desktop resize.  This mimics the behavior of
the TigerVNC Server.

Fixes #165