Server: Fix flickering w/ UltraVNC Viewer
Regression caused by 5cb50b8f8732d0d79689e207fd67bc9692c38e69

The UltraVNC Viewer seems to not send incremental FBURs, so we need to
separate the pending resize logic for "normal" and extended desktop size
messages and trigger a pending resize for the "normal" flavor only in
response to a server-side desktop resize.  This mimics the behavior of
the TigerVNC Server.

Fixes #165