Server: Clip cl->ifRegion to screen if IC enabled
... because, if interframe comparison is enabled, the interframe region
will become the new update region.  Failing to do this caused the
UltraVNC Viewer to crash, which has been the case since
ecb74b675db3a6eb5e3efb968a74f72fa796eeea.  This commit is likely the
real fix for the segfault that ecb74b675db3a6eb5e3efb968a74f72fa796eeea
was supposed to fix.

NOTE: The underlying problem is that the viewer may send us a few FBURs
between the time that the server changes the desktop size and the viewer
receives and responds to the new size.  I don't know of another way to
work around that, apart from clipping the FBUs to the screen boundary.