Linux Helper: Fix EID button synchronization issue
The EID event handler (Viewport.processExtInputEvent(), which is
implemented using JNI in the TurboVNC Helper) is called by the Java
TurboVNC Viewer's mouse button and motion callbacks, which are called by
the Swing event dispatch thread.  The EID event handler basically just
walks the X11 event queue and detects whether there are any EID motion
or button events in it.  If an EID motion event is detected, then it is
likely that more EID motion events are already in the X11 event queue,
so the EID event handler coalesces any outstanding EID motion events
into one in order to avoid pointer lag.  However, this behavior is
incorrect for EID button events, since coalescing them leads to a loss
of synchronization between the physical EID button state on the client
and the virtual EID button state within TurboVNC.

This commit modifies the EID event handler such that it stops coalescing
events as soon as it detects an EID button event.

Fixes #176