Server: Eliminate GnuTLS support
Since the OpenSSL wrapper now supports ECDH, there is no compelling
reason to use the GnuTLS wrapper anymore.  It was easy to include
support for both libraries in TurboVNC 2.1, Since both wrappers
originally came from LibVNCServer.  In 2019, however, there are no
longer any technical advantages to the GnuTLS wrapper, it basically
hasn't been tested since it was introduced, and it represents a
maintenance burden.  Furthermore, the experimental WebSockets proxy that
is currently under development would have been difficult to implement
using GnuTLS (due to the lack of an equivalent to SSL_peek().)