Travis: Use xcode8.3 image
xcode7.3 is based on El Capitan, which is EOL, and Homebrew no longer
provides El Cap bottles (pre-compiled binaries.)  Thus, Homebrew was
trying to build the packages we need from source, which slowed the Mac
CI builds.  Upgrading to Xcode 8.3.x future-proofs the CI builds against
the inevitable discontinuation of the xcode7.3 image.