Viewer: Use whole area of mixed-res/offset screens
... in full-screen mode with multi-screen spanning, automatic desktop
resizing, and Xinerama enabled.

Whereas preventing the full-screen multi-screen window from extending
beyond the unshared boundaries of any client-side physical screen is the
right thing to do in most cases, it's the wrong thing to do if the
server's virtual multi-screen layout is being set based on the client's
physical multi-screen layout.  In that case, the full-screen
multi-screen window should extend to the bounding box of all physical
screens, in order to avoid unusable screen real estate if the screens
have mismatched resolutions or are offset.

Fixes #203