SessMgr: Make 'vncserver -sessionlist' extensible
This commit modifies the SessionList() function in vncserver so that it
prepends the session list with a session count and field size and
separates fields with tabs instead of spaces.  The field size is
currently 1 but can be increased later, if we decide to add more session
information to the -sessionlist output.  That session information can
now easily contain spaces as well, since we're using tabs as field
separators.  This commit also modifies the Session Manager parser so
that it reads and verifies the session count and field size, as well as
(for future-proofness) ignores any fields other than the ones it needs.

Refer to #148