Java: Fix nuking of None,VNC sec types by opts dlg
When setting the state of the check boxes in the TurboVNC Viewer Options
dialog, we need to select the "Session encryption:None" check box if the
VNC or None security type is enabled along with an encrypted VeNCrypt
security type.  Otherwise, VNC and None will not be enabled if the user
clicks "OK", because the dialog enables VNC and None only if the
"Session encryption:None" check box is selected or the "VeNCrypt" check
box is not selected.

Note that the correspondence between security types and check boxes in
the TurboVNC Viewer Options dialog is still not 1:1.  That will be
addressed in 3.0.