UserPreferences.java: Remove unused/incorrect code
The code in question was introduced in
23a5f12431314e37b93e491f76ce9e5c27ba3d71.  It would only ever be invoked
if the viewer stored a Preferences key, other than the enc* and sec*
keys, that didn't match a configuration parameter name.  However, the
viewer never does that.  If the viewer did do that, then a logic error
in the code in question would cause the key's value to overwrite the
value of the X509CA and X509CRL parameters, which is incorrect behavior.