Server/-nevershared: Gracefully reject new conns.
With -nevershared and -dontdisconnect, we can reject a new connection by
sending an authentication failure message, which will show up in a
viewer dialog, rather than unceremoniously dropping the new connection.
This emulates the behavior of the TigerVNC Server.  Note that it is
still necessary to unceremoniously drop the new connection if
-dontdisconnect is specified without -nevershared and if a viewer does
not request a shared connection, because the server won't find out
whether a viewer requests a shared connection until the client init
message is received.