Build: Fix FreeBSD build errs w/ sys TurboJPEG lib
... and TVNC_SYSTEMX11=0

To build TurboVNC using the system-supplied TurboJPEG API library on
FreeBSD, one would pass

    -DTJPEG_INCLUDE_DIR=/usr/local/include \
    -DTJPEG_LIBRARY=/usr/local/lib/libturbojpeg.[a|so]

to CMake, but this caused /usr/local/include to be included before all
other directories, including the X11 header directories in the TurboVNC
source tree.  Since the system-supplied X11 headers are also under
/usr/local/include, this effectively caused the in-tree X.org
infrastructure to be built against the newer system-supplied X.org
headers, which led to numerous "conflicting types" errors.

This seems to be a FreeBSD-specific problem, because that is the only
O/S that supplies TurboJPEG headers in a directory other than
/usr/include.  (Because /usr/include is an "implicit" include directory,
CMake does not explicitly add it to the compiler command line.)

Fixes #250