vncserver: also parse sec. types from config file
Due to an oversight, $authTypeVNC (which determines whether to create
the VNC password file if it doesn't exist and whether to pass -rfbauth
to Xvnc), $authTypeOTP (which determines whether to ignore a user's
request to generate an initial OTP), and $encTypeX509 (which determines
whether to pass -x509cert and -x509key to Xvnc) were only set based on
the -securitytypes argument.  This commit modifies vncserver so that it
sets those variables based on the value of the $securityTypes variable
in turbovncserver.conf as well.