Win: Fix build with Windows 10 SDK
Apparently GetProcessId() is not actually available except on Windows XP
and later, but prior versions of the Windows SDK did not properly
enforce that.  It is possible that the Windows TurboVNC Viewer has not
worked on Windows 2000 since 7c70863ffd7f799dd94cacdb76a637a0f463357d
(TurboVNC 2.2.5).  Fortunately, we can use GetCurrentProcessId(), which
is available on Windows 2000, rather than
GetProcessId(GetCurrentProcess()) (not that we care about Windows 2000
support anymore-- this is just the easiest way to fix the build.)