Java: Properly handle KeyEvent.VK_SEPARATOR
On Un*x systems, if a particular key map (German, for instance)
generates a comma rather than a period for the numeric keypad decimal
key, Java generates a VK_SEPARATOR key code.  We need to map this to
XK_KP_Separator (0xFFAC) in order to emulate the behavior of Linux.