Server: show cursor to clients that don't control it
When multiple clients are connected and one of them
is moving the cursor, other clients should see an
image of the cursor moving. So it must be sent to
them via framebuffer updates. The client that moves
the cursor must not get it via framebuffer updates
(if it supports the local cursor of course).

Intended for collaborative sessions.

It seems that RealVNC has this feature.

Related to: https://github.com/novnc/noVNC/issues/974