Server: Always show ptr. movement to collaborators
When multiple viewers are sharing a single TurboVNC session, ordinarily
it would be necessary for a viewer to disable remote cursor shape
updates in order to see the pointer movements initiated by other
viewers.  This commit automates that by using server-side cursor
rendering (drawing the cursor into the framebuffer, not using remote
cursor shape updates) to send cursor updates to all viewers that aren't
moving the pointer ("collaborators.")  The viewer that is moving the
pointer (the "pointer owner") receives remote cursor shape updates while
it is moving the pointer, if it has elected to receive them.

This mimics the behavior of RealVNC.

Related to:
https://github.com/novnc/noVNC/issues/974

Based on:
https://github.com/Anatoscope/turbovnc/commit/6f6cf066baf32f9ead9d509c22b930bcf43a0f91

Closes #263