xstartup: Explicitly create session bus instance
On RHEL/CentOS/Fedora and Solaris 11, xinitrc and Xsession do exactly
this if the DBUS_SESSION_BUS_ADDRESS environment variable is empty
(which is why xstartup.turbovnc emptied it, historically speaking), so
this change has no effect on those systems.

On SuSE/openSUSE and FreeBSD, the window manager startup also
automatically creates a new session bus instance when
DBUS_SESSION_BUS_ADDRESS is empty, although the mechanism behind that
seems to be opaque (that is, it doesn't appear to be implemented in the
startup scripts.)

On Debian-compatible systems, this change prevents
/etc/X11/Xsession.d/20dbus_xdg-runtime from setting
DBUS_SESSION_BUS_ADDRESS to the per-user default value, which would
eliminate the ability to start multiple instances of the same window
manager under the same user account.  (Previously, it was necessary to
add ":noxsession" to the window manager string on Ubuntu 18+ in order to
start multiple instances of the same window manager.)