Helper: De-duplicate first touch & pointer events
In order to avoid duplicate pointer events for the first touch, we need
to handle pointer-emulation multitouch events in the Event Dispatch
Thread, as we do drawing tablet events.  The multitouch listener thread
continues to handle multitouch events for the second and subsequent
touches.

NOTE: I experimented with this prior to pushing the previous commit, but
I wasn't able to make it work at the time because I didn't realize that
the cookie was already retrieved within the context of the predicate
function.