Java: Don't recreate viewport if desktop == null
CConn.getOptions() does not always check whether desktop != null before
setting recreate = true (specifically, it does not do this if view-only
mode is toggled.)  This led to a NullPointerException if the TurboVNC
Viewer Options dialog was popped up from the "New TurboVNC Connection"
dialog (prior to connecting to a VNC server) and view-only mode was
toggled without also toggling full-screen mode.  In the general case,
recreateViewport() will throw a NPE if desktop == null, so the most
robust fix is to immediately return from recreateViewport() if that
condition is true.