Viewer: Menu options & hotkeys for zooming in/out
(when non-automatic desktop scaling is enabled)

Note that, unfortunately, there is no straightforward way to use '+' as
an accelerator in Swing, since that character is usually not located on
the primary keyboard layer.  Because the key codes in Swing are only
valid for a U.S. keyboard layout, there is not enough information in the
KeyEvent to determine, with any degree of reliability, whether
Ctrl-Alt-Shift + or Ctrl-Alt-Shift - has been pressed.  Thus, the new
implementation of this feature uses the 8 and 9 keys for, respectively,
zooming out and in instead of the - and + keys.  This provides a similar
feel to the old implementation when using U.S. keyboards, since the
8 and 9 keys are also adjacent to the 0 key.

Refer to #135