Server: Work around cong control algorithm misfit
As implemented in the TigerVNC Server, the overhauled congestion control
algorithms will, if the ETA is -1 (which would be the case if baseRTT is
(unsigned)-1, which would be the case after baseRTT is reset due to an
idle connection), report congestion without setting the congestion
timer.  TigerVNC can get away with that because it has a "frame timer"
that will deliver the congestion-deferred updates at the next frame
boundary, but in the case of TurboVNC, deferring updates due to
congestion without setting the congestion timer results in the updates
not being delivered in a timely manner.

Our implementation now treats ETA <= 0 as uncongested, since it doesn't
make much sense to set a 1 ms congestion timer if the uncongested ETA is
0.