Mac: Fix/work around more full-screen issues
Because of the Java issue described in
84c6e8cf62362871052235e3e0cf75b3809db13e, we must reconfigure rather
than recreate the viewport in CConn.getOptions() when running on macOS.
Since reconfigureViewport() is not guaranteed to invoke the viewport's
componentResized() method, CConn.getOptions() needs to ensure that the
appropriate scrollbar policy is set and the viewport is repainted after
it has been reconfigured.  Failing to do so caused two issues:

1. Increasing the scaling factor such that the desktop image no longer
fit on the screen did not cause scrollbars to appear.

2. Decreasing the scaling factor such that the desktop image was smaller
than the screen caused the old desktop image to appear in the unused
screen area around the new desktop image.

Since both of these issues were obscured by one of the issues fixed in
84c6e8cf62362871052235e3e0cf75b3809db13e, they do not merit an
additional change log entry.

This commit also fixes a logic error in CConn.getOptions() whereby
the old and new desktop size could never be equal because their screen
IDs differed.  This caused firstUpdate to always be true, which
triggered a full framebuffer update request even if one was not needed.