vncserver.man: -fg affects X startup, not Xvnc
The -fg option was introduced in TurboVNC 0.3.2, and at the time, there
was no distinction between running Xvnc and running the X startup
script.  The -noxstartup option, which was introduced in TurboVNC 1.1.x,
changed that, but due to an oversight, I neglected to modify the
vncserver man page accordingly.

Refer to #318