Viewer: Fix ExtSSH with TurboVNC Session Manager
Although this is not a supported or documented configuration yet (refer
to #148), there are some workflows for which it is useful.

Caveats:

When using an external SSH client with the TurboVNC Session Manager, the
built-in SSH client will still be used for the Session Manager itself.
The external SSH client will only be used when connecting to a session.
Thus, two separate SSH sessions will remain open (one from the built-in
SSH client and one from the external SSH client) as long as the
connection is active.  Also, if public key authentication is not used,
then it will be necessary to authenticate twice with the SSH server.

Fixes #321