JSch: Adapt OpenSSH v1 key support to our impl.
Because of 58986b73e33663ca26b61e75095d467cff8a673f and
ed506508845dfbb38fb2b30ccb59e69a40b51681, our implementation of JSch
attempts to match the public key blob of a specified private key with
a public key blob provided by the SSH agent.  We do this in order to
avoid prompting the user for an SSH key passphrase for a specified
private key if the same private key is already available in decrypted
form from the agent.  For reasons I don't yet fully understand (perhaps
I missed a commit in https://github.com/mwiede/jsch that should have
been merged), this doesn't work with the public key blob provided in
OpenSSH v1 keys, so as with PKCS keys, we have to rely on the associated
.pub file in order to implement password-less SSH authentication.  That
is a known limitation of our implementation of JSch relative to OpenSSH.